fixed exception
git-svn-id: http://svn.code.sf.net/p/tango-cs/code/gui/taurus/trunk@21667 4e9c00fd-8f2e-0410-aa12-93ce3db5e235

+++ lib/taurus/core/evaluation/ipap_example.py
@@ -64,7 +64,7 @@ class IcepapDriverParam(EvaluationDevice):
         
         try:
             value = self.ipap.readParameter(axis, param)
-            return double(value)
+            return float(value)
         except:
             return value
